+++ src/data.rs
@@ -75,45 +75,6 @@ impl Color {
         Color::rgba(0, 0, 0, 255)
     }
 
-    /// NOTE: While this function was used all the time in the
-    /// old framebuffer-based Realms library, it is rarely used
-    /// now that Realms uses opengl. You are welcome to use it,
-    /// however it will rarely be updated and may be unstable.
-    ///
-    /// This function mutates `self` to represent the color
-    /// result if the `other` color was 'added on top of'
-    /// the existing `self` color.
-    ///
-    /// If `other` is opaque, `self` becomes `other` (in
-    /// other words, the object is set to the value of
-    /// `other`).
-    ///
-    /// If `other` is *not* opaque, `self` is set to the
-    /// original value of self, but blended with `other`.
-    /// The weight of each color (how much it affects the
-    /// resulting color) is determined by the `alpha`
-    /// channel of `other`. If `other`'s alpha channel is
-    /// lower (more transparent), the result will be more
-    /// similar to the original `self`. If `other`'s alpha
-    /// channel is higher (more opaque), the result will be
-    /// more similar to `other`.
-    #[deprecated(since = "1.1.1", note = "no longer needed now that Realms uses opengl")]
-    #[expect(clippy::as_conversions, reason = "while `as` can crash, given the inputs, it is unlikely to; `as` is also necessary here")]
-    #[expect(clippy::suboptimal_flops, reason = "add_layer function no longer maintained")]
-    #[expect(clippy::cast_possible_truncation, reason = "no other feasible way to convert f64->u8 (other than `as`)")]
-    #[expect(clippy::cast_sign_loss, reason = "no other feasible way to convert f64->u8 (other than `as`)")]
-    #[inline]
-    pub fn add_layer(&mut self, other: Color) {
-        if other.a == 255 {
-            *self = other;
-        }
-        let other_alpha = f64::from(other.a) / 255.0_f64;
-        let old_alpha = 1.0_f64 - other_alpha;
-        self.r = (f64::from(self.r) * old_alpha + f64::from(other.r) * other_alpha) as u8;
-        self.g = (f64::from(self.g) * old_alpha + f64::from(other.g) * other_alpha) as u8;
-        self.b = (f64::from(self.b) * old_alpha + f64::from(other.b) * other_alpha) as u8;
-    }
-
     /// Convert a Realms `Color` into a tuple of 4 f32s, for use by opengl
     /// and (optionally) custom shaders and `VertexBuffer`s.
     #[inline]