+++ CHANGELOG.md
@@ -24,5 +24,14 @@
 
 - Add `VertexBuffer::set_layout` method
   - [https://docs.rs/realms/1.2.3/realms/vertex/struct.VertexBuffer.html#method.set_layout]
+- Update examples to use `VertexBuffer::set_layout` method
 - No breaking changes
+
+# 1.3.3 -> 2.3.3 (major, breaking)
+
+- Fixed some lints in `Shader::load_str` method
+- `Shader::load_str` function now takes `source` as `&str` instead of `String`
+  -- BREAKING
+- Deleted unused function `Color::add_layer` -- BREAKING
+- Update examples to use new `Shader::load_str`